Refactor postgres table lib and apply helper tests
