+++ src/lib/table/postgres/helpers.js
@@ -1,8 +1,7 @@
 'use strict';
 
-const config = require('../../../config');
-const postgres = require('../postgres');
-const { key2Field } = require('./helpers');
+const { key2Field } = require('../helpers');
+const moment = require('moment');
 
 require('functional-augments');
 
@@ -13,10 +12,17 @@ const term2Sql = t => {
 
   if (v instanceof RegExp) {
     return 'anglicise(' + field + ')' + ' ~ anglicise(${' + k + '})';
+  } else if (v instanceof Date || v instanceof moment) {
+    return field + ' = ${' + k + '}';
   } else if (v instanceof Object) {
     const operators = [ '>', '<', '>=', '<=' ];
 
-    return v.filter((v2, k2) => operators.includes(k)).map((v2, k2) => field + ' ' + k2 + ' ${' + k + '.' + k2 + '}');
+    const r = Object.values(v
+                            .filter((v2, k2) => operators.includes(k2))
+                            .map((v2, k2) => field + ' ' + k2 + ' ${' + k + '.' + k2 + '}')
+                           ).join('\n  AND ');
+
+    return r;
   } else {
     return field + ' = ${' + k + '}';
   }
@@ -47,23 +53,6 @@ const processTerms = terms => {
   };
 };
 
-module.exports = (tableName, fields, idField) => {
-  const select = 'SELECT ${_fields:name}';
-  const from = 'FROM ${_table:name}';
-  const limit = 'LIMIT ' + config.maxRecords;
-  const names = {
-    _id: idField,
-    _fields: fields,
-    _table: tableName
-  };
-
-  return {
-    read: id => postgres.oneOrNone([select, from, 'WHERE ${_id:name} = ${id}', limit].join('\n'), Object.assign({id: id}, names)),
-    search: terms => {
-      const processedTerms = processTerms(terms);
-      const sql = [select, from, processedTerms.sql, limit].join('\n');
-
-      return postgres.any(sql, Object.assign(processedTerms.params, names));
-    }
-  };
+module.exports = {
+  processTerms: processTerms
 };
